Refactored light/dark button code
Redid the HTML for the button itself for the light/dark exercise just to start
fresh, as well as rewrote and re-defined different attributes within the
internal CSS. Finally, refactored the JS for the light/dark button so that
it looks cleaner, less redundant, and uses new/clearer variable names. This
new version of the light/dark button uses CSS variables that adopt values
based on whether the .dark attribute is toggled onto the body element, which
also therefore changes the textContent of the button because of the if-else
statement tied to the appearance of .dark on the body. This new, refactored
version disregards using filter: invert(100%) because it altered the color
and appearance of the existing buttons. However, this code still keeps a seperate
CSS attribute that targets buttons and images incase one were to add an
invert filter again.

Fixes-Bug: Adds filter:invert(100%) prevention from targeting images (not needed
with refactored code)
Implements-Requirement: n/a